+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -23,6 +23,7 @@ import {
 import { createMapPlugin, MAP_TYPE } from 'wix-rich-content-plugin-map';
 import { createFileUploadPlugin, FILE_UPLOAD_TYPE } from 'wix-rich-content-plugin-file-upload';
 import { createTextColorPlugin, TEXT_COLOR_TYPE } from 'wix-rich-content-plugin-text-color';
+import createBlockDndPlugin from 'draft-js-drag-n-drop-plugin';
 import Highlighter from 'react-highlight-words';
 import casual from 'casual-browserify';
 
@@ -74,6 +75,7 @@ export const editorPlugins = [
   createMapPlugin,
   createFileUploadPlugin,
   createTextColorPlugin,
+  createBlockDndPlugin,
 ];
 
 const themeColors = {
@@ -180,7 +182,8 @@ export const config = {
         'JWS.eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6Im5FUXljQzlOIn0.eyJpYXQiOjE1Njc1MjY3NzQsImRhdGEiOiJ7XCJ1c2VySWRcIjpcIjE5YTY0YTRjLWVlZTAtNGYxNC1iNjI3LTY3MmQ1ZjE2OGJkNFwiLFwibWV0YXNpdGVJZFwiOlwiNTM4ZmE2YzYtYzk1My00Y2RkLTg2YzQtNGI4NjlhZWNmOTgwXCJ9IiwiZXhwIjoxNTY4NzM2Mzc0fQ.n21OxIzSbqi8N3v30b6cIxMdshBnkkf2WQLWEFVXsLk',
       metaSiteId: '538fa6c6-c953-4cdd-86c4-4b869aecf980',
       mediaRoot: 'some-mediaRoot',
-    },
+    },    
+    onImageEditorOpen: () => console.log('Media Studio Launched'),
     // toolbar: {
     //   icons: {
     //     Image: MyCustomIcon, // insert plugin icon
@@ -317,6 +320,28 @@ export const config = {
         console.log('consumer uploaded ', videoToUpload);
       }, 500);
     },
+    // handleFileUpload: (file, updateEntity, removeEntity) => {
+    //   console.log('consumer wants to upload custom video', file);
+    //   const videoWithAbsoluteUrl = {
+    //     url: 'http://mirrors.standaloneinstaller.com/video-sample/jellyfish-25-mbps-hd-hevc.mp4',
+    //   };
+    //   const videoWithRelativeUrl = {
+    //     pathname: 'video/441c23_84f5c058e5e4479ab9e626cd5560a21b/file',
+    //     thumbnail: {
+    //       pathname: 'media/441c23_84f5c058e5e4479ab9e626cd5560a21bf000.jpg',
+    //       height: 1080,
+    //       width: 1920,
+    //     },
+    //   };
+    //   // You can provide either absolute or relative URL.
+    //   // If relative URL is provided, a function 'getVideoUrl' will be invoked to form a full URL.
+    //   const videoToUpload = videoWithAbsoluteUrl;
+    //   setTimeout(() => {
+    //     updateEntity({ data: videoToUpload });
+    //     //updateEntity({ error: { msg: 'Upload Failed' } });
+    //     console.log('consumer uploaded ', videoToUpload);
+    //   }, 500);
+    // },
     enableCustomUploadOnMobile: true,
     // Function is invoked when rendering video which has relative URL.
     // You should take the pathname and form a full URL.