+++ packages/plugin-image/web/src/toolbar/createToolbar.js
@@ -17,7 +17,6 @@ export default function createToolbar({
       relValue,
       uiSettings,
       isMobile,
-      imageEditorWixSettings: settings.imageEditorWixSettings,
       settings,
     }),
     InsertButtons: createInsertButtons({ helpers, t, settings }),