+++ packages/plugin-image/web/src/toolbar/inline-buttons.jsx
@@ -5,20 +5,13 @@ import { MediaReplaceIcon, ImageEditorIcon } from '../icons';
 
 const removeEmpty = list => list.filter(item => !!item);
 
-export default ({
-  t,
-  anchorTarget,
-  relValue,
-  uiSettings,
-  isMobile,
-  imageEditorWixSettings,
-  settings,
-}) => {
+export default ({ t, anchorTarget, relValue, uiSettings, isMobile, settings = {} }) => {
   const icons = get(settings, 'toolbar.icons', {});
   const modalStyles = getModalStyles({ isMobile });
   const imageEditorStyles = getModalStyles({
     customStyles: { content: { maxWidth: '100%', background: 'transparent' } },
   });
+  const { imageEditorWixSettings, onImageEditorOpen } = settings;
   const imageEditorButton = imageEditorWixSettings
     ? {
         keyName: 'imageEditor',
@@ -28,6 +21,7 @@ export default ({
         modalStyles: imageEditorStyles,
         t,
         imageEditorWixSettings,
+        onImageEditorOpen,
         mobile: false,
         tooltipTextKey: 'ImageEditorButton_Tooltip',
         mapComponentDataToButtonProps: componentData => ({